requirements.txt
streamlit
openai
langchain
langchain_openai
python-dotenv